refactor every install.sh and main.sh

+++ bash/06_jq/install.sh
@@ -3,16 +3,48 @@
 __rayrc_install() {
     __rayrc_module_common_setup
 
-    if [[ ${__rayrc_facts_os_type} == "linux" ]]; then
-        __rayrc_github_downloader "stedolan/jq" "${__rayrc_data_dir}/jq" "linux64"
+    if [[ "${__rayrc_facts_os_type}" == "linux" ]]; then
+        if uname -m | grep -E -q "arm|aarch" >&/dev/null; then
+            if [[ ! "${__rayrc_package_manager}" =~ ^[[:space:]]*$ ]]; then
+                ${__rayrc_package_manager} update -y >&/dev/null
+                ${__rayrc_package_manager} install -y jq >&/dev/null
+            fi
+        elif uname -m | grep -E -q "64" >&/dev/null; then
+            __rayrc_github_downloader \
+                "stedolan/jq" "${__rayrc_data_dir}/jq" \
+                "linux64"
+        elif uname -m | grep -E -q "86" >&/dev/null; then
+            __rayrc_github_downloader \
+                "stedolan/jq" "${__rayrc_data_dir}/jq" \
+                "linux32"
+        else
+            echo ".rayrc: unsupported cpu architecture for downloading exa.."
+            return 8
+        fi
+    elif [[ "${__rayrc_facts_os_type}" == "macos" ]]; then
+        if uname -m | grep -E -q "86|ia64" >&/dev/null; then
+            __rayrc_github_downloader \
+                "stedolan/jq" "${__rayrc_data_dir}/jq" \
+                "osx"
+        elif uname -m | grep -E -q "arm|aarch" >&/dev/null; then
+            if [[ ! "${__rayrc_package_manager}" =~ ^[[:space:]]*$ ]]; then
+                ${__rayrc_package_manager} update -y >&/dev/null
+                ${__rayrc_package_manager} install -y jq >&/dev/null
+            fi
+        else
+            echo ".rayrc: unsupported cpu architecture for downloading exa.."
+            return 8
+        fi
     else
-        __rayrc_github_downloader "stedolan/jq" "${__rayrc_data_dir}/jq" "osx"
+        echo ".rayrc: unsupported os for downloading exa.."
+        return 8
     fi
 
-    mv -f "${__rayrc_data_dir}/jq" "${__rayrc_bin_dir}"
-    chmod -R 755 "${__rayrc_bin_dir}"
-    rm -rf "${__rayrc_data_dir}/jq"*
-
+    if [[ -f "${__rayrc_data_dir}/jq" ]]; then
+        mv -f "${__rayrc_data_dir}/jq" "${__rayrc_bin_dir}"
+        chmod -R 755 "${__rayrc_bin_dir}"
+        rm -rf "${__rayrc_data_dir}/jq"*
+    fi
 }
 
 __rayrc_install
